For #24847 - Part 2: Refactor AutofillFragmentStore into the autofill package
