+++ app/src/main/java/org/mozilla/fenix/settings/autofill/AutofillFragmentStore.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.settings.creditcards
+package org.mozilla.fenix.settings.autofill
 
 import mozilla.components.concept.storage.CreditCard
 import mozilla.components.lib.state.Action
