+++ app/src/test/java/org/mozilla/fenix/settings/autofill/AutofillFragmentStoreTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.settings.creditcards
+package org.mozilla.fenix.settings.autofill
 
 import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
